updating ontology classes to ontology file
